requirements.txt
torch
torchvision
numpy
pandas
matplotlib
pillow
tqdm
jupyter
